Finished Renaming android and iOS shared projects
