+++ Sensus.Android.Shared/Concurrent/MainConcurrent.cs
@@ -12,32 +12,42 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Foundation;
 using System;
 using System.Threading;
 using Sensus.Dispose;
 using Sensus.Concurrent;
-using Xamarin.Forms;
+using Android.OS;
 
-namespace Sensus.iOS.Concurrent
+namespace Sensus.Android.Concurrent
 {
+    /// <remarks>
+    /// Device.BeginInvokeOnMainThread invokes off the activity. Sensus does not always have an activity, so we create a handler bound to the main thread's looper instead.
+    /// </remarks>
     public class MainConcurrent : Disposable, IConcurrent
     {
         private readonly int? _waitTime;
 
+        #region Fields
+        private readonly Handler _handler;
+        #endregion
+
+        #region Constructor
         public MainConcurrent(int? waitTime = null)
         {
             _waitTime = waitTime;
+            _handler = new Handler(Looper.MainLooper);
         }
+        #endregion
 
+        #region Public Methods
         public void ExecuteThreadSafe(Action action)
         {
             if (action == null)
             {
-                throw new ArgumentNullException();
+                throw new ArgumentNullException(nameof(action));
             }
 
-            if (NSThread.IsMain)
+            if (_handler.Looper.IsCurrentThread)
             {
                 action();
             }
@@ -45,7 +55,7 @@ namespace Sensus.iOS.Concurrent
             {
                 var runWait = new ManualResetEvent(false);
 
-                NSThread.MainThread.BeginInvokeOnMainThread(() =>
+                _handler.Post(() =>
                 {
                     try
                     {
@@ -71,16 +81,16 @@ namespace Sensus.iOS.Concurrent
                 throw new ArgumentNullException(nameof(func));
             }
 
-            if (NSThread.IsMain)
+            if (_handler.Looper.IsCurrentThread)
             {
                 return func();
             }
             else
             {
-                var result = default(T);
                 var runWait = new ManualResetEvent(false);
+                var result = default(T);
 
-                NSThread.MainThread.BeginInvokeOnMainThread(() =>
+                _handler.Post(() =>
                 {
                     try
                     {
@@ -100,5 +110,17 @@ namespace Sensus.iOS.Concurrent
                 return result;
             }
         }
+
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing)
+            {
+                _handler.Dispose();
+            }
+
+            base.Dispose(disposing);
+        }
+
+        #endregion
     }
 }
