+++ Sensus.Android.Shared/Context/AndroidSensusContext.cs
