+++ Sensus.Android/Sensus.Android.csproj
@@ -356,6 +356,6 @@ Store. It is possible to run and debug this project on physical and virtual Andr
     <Folder Include="Callbacks\" />
   </ItemGroup>
   <Import Project="..\Sensus.Shared\Sensus.Shared.projitems" Label="Shared" />
-  <Import Project="..\Sensus.Shared.Android\Sensus.Shared.Android.projitems" Label="Shared" />
+  <Import Project="..\Sensus.Android.Shared\Sensus.Android.Shared.projitems" Label="Shared" />
   <Import Project="..\packages\Xamarin.Forms.2.3.2.127\build\portable-win+net45+wp80+win81+wpa81+MonoAndroid10+MonoTouch10+Xamarin.iOS10\Xamarin.Forms.targets" Condition="Exists('..\packages\Xamarin.Forms.2.3.2.127\build\portable-win+net45+wp80+win81+wpa81+MonoAndroid10+MonoTouch10+Xamarin.iOS10\Xamarin.Forms.targets')" />
 </Project>
