--- Sensus.Shared.iOS/Sensus.Shared.iOS.projitems
@@ -1,47 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>ca15dd76-c484-4a56-a03f-9ef242294d55</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>Sensus.iOS</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Concurrent\MainConcurrent.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Context\iOSSensusContext.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Exceptions\iOSInsightsInitializer.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\Apps\iOSFacebookProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\Context\iOSSoundProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\Device\iOSBatteryProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\Location\iOSCompassProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\Movement\iOSAccelerometerProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\Network\iOSPollingWlanProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\User\Health\iOSHealthKitBiologicalSexProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\User\Health\iOSHealthKitBirthdateProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\User\Health\iOSHealthKitBloodTypeProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\User\Health\iOSHealthKitHeightProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\User\Health\iOSHealthKitProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Probes\User\Health\iOSHealthKitSamplingProbe.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\iOSCallbackScheduler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\UILocalNotifications\UILocalNotificationCallbackScheduler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\UILocalNotifications\UILocalNotificationNotifier.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\UNUserNotifications\UNUserNotificationCallbackScheduler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\UNUserNotifications\UNUserNotificationNotifier.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\IiOSCallbackScheduler.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\UILocalNotifications\IUILocalNotificationNotifier.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\UNUserNotifications\UNUserNotificationDelegate.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\UNUserNotifications\IUNUserNotificationNotifier.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\IiOSNotifier.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Callbacks\iOSNotifier.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UI\SliderInputEffect.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="$(MSBuildThisFileDirectory)Probes\" />
-    <Folder Include="$(MSBuildThisFileDirectory)Callbacks\" />
-    <Folder Include="$(MSBuildThisFileDirectory)Callbacks\UILocalNotifications\" />
-    <Folder Include="$(MSBuildThisFileDirectory)Callbacks\UNUserNotifications\" />
-    <Folder Include="$(MSBuildThisFileDirectory)UI\" />
-  </ItemGroup>
-</Project>
