+++ Sensus.iOS.Shared/Callbacks/UNUserNotifications/UNUserNotificationCallbackScheduler.cs
