+++ Sensus.iOS.Shared/Callbacks/UNUserNotifications/UNUserNotificationDelegate.cs
