+++ Sensus.iOS.Shared/Probes/Device/iOSBatteryProbe.cs
