+++ Sensus.iOS.Shared/Probes/Location/iOSCompassProbe.cs
