+++ Sensus.iOS.Shared/Probes/Movement/iOSAccelerometerProbe.cs
