+++ Sensus.iOS.Shared/Probes/User/Health/iOSHealthKitBloodTypeProbe.cs
