+++ Sensus.iOS.Shared/Probes/User/Health/iOSHealthKitHeightProbe.cs
