+++ Sensus.iOS.Shared/UI/SliderInputEffect.cs
@@ -12,26 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Android.Widget;
-using Xamarin.Forms.Platform.Android;
 using Xamarin.Forms;
-using Android.Graphics.Drawables;
-using Sensus.Shared.Android.UI;
+using Xamarin.Forms.Platform.iOS;
+
+using UIKit;
+using CoreGraphics;
+
+using Sensus.iOS.UI;
 using Sensus.UI.Inputs;
 
+
+// register the slider input effect
 [assembly: ExportEffect(typeof(SliderInputEffect), SliderInput.EFFECT_RESOLUTION_EFFECT_NAME)]
 
-namespace Sensus.Shared.Android.UI
+namespace Sensus.iOS.UI
 {
-    /// <summary>
-    /// Hides the slider thumb button until the user taps the slider. This is important because displaying the thumb
-    /// button immediately can bias the user to (1) move the button away from the initial value or (2) bias the user
-    /// to select a value near the initial value (anchor-adjustment bias).
-    /// </summary>
     public class SliderInputEffect : PlatformEffect
     {
+        private UIImage _visibleThumbImage;
         private bool _userHasChangedSliderValue;
-        private Drawable _visibleThumbDrawable;
 
         public SliderInputEffect()
         {
@@ -40,30 +39,34 @@ namespace Sensus.Shared.Android.UI
 
         protected override void OnAttached()
         {
-            SeekBar nativeSeekBar = Control as SeekBar;
+            UISlider nativeSlider = Control as UISlider;
             Slider formsSlider = Element as Slider;
 
             // make the thumb image invisible if the user hasn't previously changed the slider's value
             if (!_userHasChangedSliderValue)
             {
-                _visibleThumbDrawable = nativeSeekBar.Thumb;
-                nativeSeekBar.SetThumb(new ColorDrawable(global::Android.Graphics.Color.Transparent));
+                _visibleThumbImage = nativeSlider.ThumbImage(UIControlState.Normal);
+                nativeSlider.SetThumbImage(new UIImage(), UIControlState.Normal);
             }
 
-            nativeSeekBar.ProgressChanged += (sender, e) =>
+            // listen for the user pressing on the slider. this also reports slides.
+            nativeSlider.AddGestureRecognizer(new UILongPressGestureRecognizer(pressRecognizer =>
             {
-                // make the thumb image visible if we haven't already
+                // make the thumb image visible
                 if (!_userHasChangedSliderValue)
                 {
-                    nativeSeekBar.SetThumb(_visibleThumbDrawable);
+                    nativeSlider.SetThumbImage(_visibleThumbImage, UIControlState.Normal);
                     _userHasChangedSliderValue = true;
                 }
 
-                // set the value on the forms slider control. note that SeekBars have a fixed minimum (0) and
-                // a variable maximum. rescale the SeekBar's value to be in the range desired for the forms control.
-                double percent = nativeSeekBar.Progress / (double)nativeSeekBar.Max;
+                // update the slider value - we subtract 25 from the width of the frame to get the right offset (fudgey)
+                CGPoint pointTapped = pressRecognizer.LocationInView(pressRecognizer.View);
+                float percent = (float)((pointTapped.X - nativeSlider.Frame.Location.X) / (nativeSlider.Frame.Size.Width - 25));
+                float newNativeValue = nativeSlider.MinValue + (percent * (nativeSlider.MaxValue - nativeSlider.MinValue));
+                nativeSlider.SetValue(newNativeValue, false);
                 formsSlider.Value = formsSlider.Minimum + (percent * (formsSlider.Maximum - formsSlider.Minimum));
-            };
+            })
+            { MinimumPressDuration = 0 });
         }
 
         protected override void OnDetached()
